+++ types/index.d.ts
@@ -263,6 +263,8 @@ declare module '@aller/openid-connect' {
 	 *  [...]
 	 * })
 	 * ```
+	 *
+	 * Set to `autodetect` if trust proxy is enabled. Base URL will be built from headers `x-forwarded-proto` and `x-forwarded-host`
 	 */
 	baseURL?: string;