+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Breaking
 - default `clientAuthMethod` to `client_secret_basic` if client secret is configured
 
+### Additions
+- configured `baseURL` accepts the string `autodetect` that will build `baseURL` from `req.protocol` and `req.host`. Simplifies caching of OpenID Clients since they are cached with configuration object ref as key
+
 ## v0.0.3 (2026-03-16)
 
 Refactor cookie handling that unfortunately breaks v0.0.1-2 created cookies. Latest [jose](https://www.npmjs.com/package/jose) uses nodejs builtin [Web Crypto API](https://nodejs.org/docs/latest-v22.x/api/webcrypto.html) ([mdn](https://developer.mozilla.org/en-US/docs/Web/API/Web_Crypto_API)).